move Static -> Typed and add device type annotation to tensor (wip)
